+++ src/main/kotlin/com/insyncwithfoo/pyright/configuration/application/Configurable.kt
@@ -8,7 +8,6 @@ internal class Configurable : PyrightConfigurable<Configurations>() {
     
     override val service = ConfigurationService.getInstance()
     override val state = service.state.copy()
-    override val originalState = state.copy()
     
     override val panel by lazy { configurationPanel(state) }